# svc-candidate-api/src/main/resources/openapi/yaml/api.yaml
openapi: 3.0.3
info:
  title: Library API - JobBeacon Candidate Registration API
  description: |-
    This API provides basic operation for creation, modification and deletion of JobBeacon AI candidates.


    Some useful links:
    - [The library repository](https://github.com/andrewkandzuba/openexchange-jobbeacon-ai)
    - [The source API definition for the library](https://github.com/andrewkandzuba/openexchange-jobbeacon-ai/svc-candidate-api/blob/main/src/main/resources/openapi/yaml/api.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: andrey.kandzuba@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: '0.0.1'

externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io

servers:
  - url: '{protocol}:/api/v1'
    variables:
      protocol:
       enum:
         - http
         - https
       default: http

tags:
  - name: Candidate
    description: Operations about candidate
    externalDocs:
      description: Find out more
      url: http://swagger.io

paths:
  /candidates:
    post:
      tags:
        - Candidate
      summary: Register a new candidate.
      description: Register new candidate in platform
      operationId: registerCandidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Registration'
            example:
              $ref: '#/components/examples/registration'
      responses:
        '201':
          description: Created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
              example:
                $ref: '#/components/examples/candidate'
        '400':
          description: 'something went wrong!'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: You are not authorized to view the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accessing the resource you were trying to reach is forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Candidate
      summary: List candidates.
      description: Retrieve all candidates from platform.
      operationId: listCandidates.
      responses:
        '200':
          description: 'Read successfully'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Candidate'
        '400':
          description: 'something went wrong!'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: You are not authorized to view the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accessing the resource you were trying to reach is forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /candidates/{id}:
    get:
      tags:
        - Candidate
      summary: Retrieve a candidate profile.
      description: Returns a single record by candidate id
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Candidate is found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
              example:
                $ref: '#/components/examples/candidate'
        '400':
          description: 'something went wrong!'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: You are not authorized to view the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accessing the resource you were trying to reach is forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The resource you were trying to reach is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Candidate
      summary: Update a candidate profile.
      description: Update a candidate profile to platform
      operationId: updateCandidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
            example:
              $ref: '#/components/examples/candidate'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
              example:
                $ref: '#/components/examples/candidate'
        '400':
          description: 'something went wrong!'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: You are not authorized to view the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Accessing the resource you were trying to reach is forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: The resource you were trying to reach is not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
    schemas:
        Error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required:
            - code
            - message
        Registration:
          required:
            - email
            - firstName
            - lastName
            - city
            - country
            - zip
          type: object
          properties:
            email:
              maxLength: 255
              minLength: 1
              type: string
              description: candidate email
            firstName:
              maxLength: 50
              minLength: 1
              type: string
              description: Candidate First Name
            lastName:
              maxLength: 50
              minLength: 1
              type: string
              description: Candidate Last Name
            city:
              maxLength: 100
              minLength: 1
              type: string
              description: Candidate City of Residential
            country:
              maxLength: 100
              minLength: 1
              type: string
              description: Candidate Country of Residential
            zip:
              maxLength: 10
              minLength: 5
              type: string
              description: Candidate Zipcode
        Candidate:
          required:
            - id
            - email
            - firstName
            - lastName
            - city
            - country
            - zip
          allOf:
            - $ref: '#/components/schemas/Registration'
            - type: object
              required:
                - id
              properties:
                id:
                  type: integer
                  format: int64
    examples:
        candidate:
          value:
            id: 1
            email: candidate@gmail.com
            firstName: John
            lastName: Smith
            country: The United States
            zip: 12345
          summary: A created candidate
        registration:
          value:
            email: candidate@gmail.com
            firstName: John
            lastName: Smith
            country: The United States
            zip: 12345
          summary: A new candidate